requirements.txt
aiofiles==0.8.0
aiohttp==3.8.4
aiolimiter==1.0.0
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
APScheduler==3.10.1
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.7.0
asyncprawcore==2.3.0
attrs==23.1.0
backports.zoneinfo==0.2.1
bitarray==2.7.3
Cerberus==1.3.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
coloredlogs==15.0.1
cytoolz==0.12.1
dateparser==1.1.8
dnspython==2.3.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
exceptiongroup==1.1.1
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
hexbytes==0.3.0
hiredis==2.2.2
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
idna==3.4
importlib-resources==5.12.0
jsonschema==4.17.3
lru-dict==1.2.0
mnemonic==0.20
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
outcome==1.2.0
pandas==2.0.1
parsimonious==0.9.0
peewee==3.16.2
pkgutil-resolve-name==1.3.10
protobuf==4.23.2
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.17
pymongo==4.3.3
pyreadline3==3.4.1
Pyrogram==2.0.104
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
redis==4.5.4
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rlp==3.0.0
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
TgCrypto==1.2.5
toolz==0.12.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
update-checker==0.18.0
urllib3==1.26.15
web3==6.4.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
